[TASK] add --symlink-name option
the option is optional, it's default value is "PhpStorm".

to specify a filename (basename) of the symlink that is representing the
directory to the just installed version within the "opt" directory after
extraction.

we had the scenario that the standard name was already used by the puppet
managed installation and therefore overwritten. with the new parameter
the user-specific name could be used so that a user herself could easily
do the EAP install.